feat: 카메라 기능 구현
simulator로 켜지지 않아서 예외처리만 해두었습니다.
